requirements.txt
asgiref==3.7.2
astroid==3.1.0
cfgv==3.4.0
dill==0.3.8
distlib==0.3.8
Django==5.0.3
djangorestframework==3.14.0
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==24.2.0
filelock==3.13.1
identify==2.5.35
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
nodeenv==1.8.0
packaging==24.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.2
psycopg2-binary==2.9.9
pytest==8.1.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
tomlkit==0.12.4
uritemplate==4.1.1
virtualenv==20.25.1
